Added automatic ADC Cut derivation
Tool can now automatically grab an ADC cut value. The way this works is that the tool finds the peak (mode) of the ADC data and tries to encapsulate 68% of the data from that point. In practice, it tends to grab around 70-80% of the data due to the bin sizes.

This feature can be manually overriden if reAdjust is set to false when running bhm.analyse(). There is currently no way to set this (in either the gui or non gui analysis) other than manually changing this value, I may add an option for this in the future.